requirements.txt
Django~=4.2
asgiref~=3.6.0
Pillow~=9.5.0
pip==23.1.2
psycopg2-binary==2.9.3
setuptools==67.7.2
sqlparse~=0.4.4
tzdata~=2023.3
djangorestframework~=3.14.0
drf-yasg~=1.21.5
pylint~=2.17.3
black~=23.3.0
pytest~=7.3.1


requests~=2.30.0